requirements.txt
# requirements.txt

unittest
re
unittest TestCase
time
selenium webdriver
selenium.webdriver.common.keys Keys
selenium.webdriver.common.by By
selenium.webdriver.support.ui WebDriverWait
selenium.webdriver.support expected_conditions
webdriver_manager.chrome ChromeDriverManager
requests
json